+++ src/smon_version.py
@@ -1,0 +1,1 @@
+SMON_VERSION = "0.5.2"